# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory Management System API
  description: |
    REST API for the Inventory Management System (IMS).

    Features include product catalog, multi-warehouse inventory, suppliers,
    purchase orders, sales orders, stock movements, reports, and audit logs.

    Real-time WebSocket endpoints (not part of this REST spec):
    - `ws://localhost:3000/ws/inventory`
    - `ws://localhost:3000/ws/products`
    - `ws://localhost:3000/ws/warehouses`
    - `ws://localhost:3000/ws/suppliers`
  version: 1.1.0
  contact:
    name: IMS API Support
servers:
  - url: /
    description: Current host (use for Swagger Try it out)
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /reports/stock-summary:
    get:
      tags: [Reports]
      summary: Get stock summary report
      description: Returns stock value breakdown by warehouse and product.
      operationId: getStockSummary
      responses:
        "200":
          description: Stock summary report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockSummaryResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /employees:
    get:
      tags: [Employees]
      summary: List employees
      operationId: listEmployees
      parameters:
        - name: role
          in: query
          schema: { type: string }
        - name: warehouse_id
          in: query
          schema: { type: integer }
        - name: active
          in: query
          schema: { type: string, enum: [true] }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
    post:
      tags: [Employees]
      summary: Create employee
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
  /employees/{id}:
    parameters:
      - $ref: "#/components/parameters/EmployeeId"
    get:
      tags: [Employees]
      summary: Get employee
      operationId: getEmployee
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
    put:
      tags: [Employees]
      summary: Update employee
      operationId: updateEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
    delete:
      tags: [Employees]
      summary: Delete employee
      operationId: deleteEmployee
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /reports/top-products:
    get:
      tags: [Reports]
      summary: Top selling products
      operationId: getTopProducts
      parameters:
        - name: limit
          in: query
          schema: { type: string, default: "10" }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopProductsResponse"
  /audit-logs:
    get:
      tags: [Reports]
      summary: Get audit logs
      description: Returns the latest 100 audit log entries.
      operationId: getAuditLogs
      parameters:
        - name: entity
          in: query
          description: Filter by entity type (e.g. Product, Order)
          schema:
            type: string
        - name: action
          in: query
          description: Filter by action (e.g. CREATE, UPDATE)
          schema:
            type: string
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  parameters:
...
    EmployeeId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: uint
...
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        text/plain:
          schema:
            type: string
            example: Failed to fetch products
  schemas:
...
    MessageResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Product deleted successfully
...
    StockSummaryItem:
      type: object
      properties:
        warehouse_name:
          type: string
        product_name:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        value:
          type: number
          format: double
    StockSummaryResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/StockSummaryItem"
            total_value:
              type: number
              format: double
            total_items:
              type: integer
    AuditLog:
      type: object
      properties:
        id:
          type: integer
          format: uint
        action:
          type: string
        entity:
          type: string
        entity_id:
          type: integer
          format: uint
        user_id:
          type: string
        details:
          type: string
        ip_address:
          type: string
        created_at:
          type: string
          format: date-time
    AuditLogListResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
...
    Employee:
      type: object
      properties:
        id: { type: integer, format: uint }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [admin, manager, staff] }
        phone: { type: string }
        warehouse_id: { type: integer, format: uint, nullable: true }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateEmployeeRequest:
      type: object
      required: [name, email]
      properties:
        name: { type: string }
        email: { type: string }
        role: { type: string }
        phone: { type: string }
        warehouse_id: { type: integer, format: uint }
        is_active: { type: boolean }
    EmployeeResponse:
      type: object
      properties:
        status: { type: string, example: success }
        data: { $ref: "#/components/schemas/Employee" }
    EmployeeListResponse:
      type: object
      properties:
        status: { type: string, example: success }
        data:
          type: array
          items: { $ref: "#/components/schemas/Employee" }
...
    TopProductsResponse:
      type: object
      properties:
        status: { type: string, example: success }
        data:
          type: array
          items:
            type: object
            properties:
              product_id: { type: integer }
              product_name: { type: string }
              sku: { type: string }
              units_sold: { type: integer }
              revenue: { type: number }